refactor(examples): share API types between client and worker
OutputType, TextLayer, and the request/result shapes of the three AI routes
were defined twice — src/api/marketingApi.ts and worker/providers/types.ts —
so a field added on one side could silently drift from the other. They now
live once in shared/apiTypes.ts (types only, worker-safe), imported by both
sides; the compiler checks the seam. The client keeps the fetch functions,
the worker keeps its provider interfaces and *Params aliases.
