+++ apps/examples/marketing-assets/src/api/marketingApi.ts
@@ -1,101 +1,37 @@
+import type {
+	ClarifyRequest,
+	ClarifyResult,
+	GenerateRequest,
+	GenerateResult,
+	PlanRequest,
+	PlanResult,
+} from '../../shared/apiTypes'
+
 /**
- * Frontend API client for the Cloudflare Worker backend.
- * Mirrors the worker's provider types so the contract stays in one shape.
+ * Frontend API client for the Cloudflare Worker backend. The request and result
+ * shapes are the shared wire contract in shared/apiTypes.ts, imported by both this
+ * client and the worker's providers.
  */
 
-export interface OutputType {
-	id: string
-	label: string
-	width: number
-	height: number
-	/** The platform this format targets, used to group the format picker. */
-	platform?: string
-}
-
-/** A piece of text the app renders deterministically over the background. */
-export interface TextLayer {
-	text: string
-	x: number
-	y: number
-	width: number
-	fontRole: 'heading' | 'body'
-	fontSize: number
-	color: string
-	align: 'left' | 'center' | 'right'
-	weight: 'normal' | 'bold'
-	scrim: boolean
-}
-
-export interface GenerateRequest {
-	prompt: string
-	brandText: string
-	outputType: OutputType
-	referenceImages: string[]
-	/** On a background edit, the current background as a data URL. */
-	currentImage?: string
-	/** On a background edit, the instruction to apply. */
-	instruction?: string
-}
-
-export interface GenerateResult {
-	imageUrl: string
-}
+export type {
+	ClarifyRequest,
+	ClarifyResult,
+	GenerateRequest,
+	GenerateResult,
+	OutputType,
+	PlanRequest,
+	PlanResult,
+	TextLayer,
+} from '../../shared/apiTypes'
 
 /** Render or edit the text-free background image. */
 export async function apiGenerate(params: GenerateRequest): Promise<GenerateResult> {
-	const response = await fetch('/api/generate', {
-		method: 'POST',
-		headers: { 'Content-Type': 'application/json' },
-		body: JSON.stringify(params),
-	})
-	if (!response.ok) {
-		const err = await response.json().catch(() => ({ error: response.statusText }))
-		throw new Error((err as { error?: string }).error ?? 'Generation failed')
-	}
-	return (await response.json()) as GenerateResult
-}
-
-export interface PlanRequest {
-	mode: 'create' | 'revise'
-	prompt: string
-	brandText: string
-	outputType: OutputType
-	image: string
-	currentLayers?: TextLayer[]
-	annotations?: string[]
-	/** A distinct messaging angle for this asset, so a batch's captions vary. */
-	captionAngle?: string
-}
-
-export interface PlanResult {
-	textLayers: TextLayer[]
-	/** Body copy shown beside the asset (the social caption), not on the image. */
-	caption: string
-	backgroundInstructions: string[]
+	return post('/api/generate', params, 'Generation failed')
 }
 
 /** Plan the text layers (and, on revise, any background edits). */
 export async function apiPlan(params: PlanRequest): Promise<PlanResult> {
-	const response = await fetch('/api/plan', {
-		method: 'POST',
-		headers: { 'Content-Type': 'application/json' },
-		body: JSON.stringify(params),
-	})
-	if (!response.ok) {
-		const err = await response.json().catch(() => ({ error: response.statusText }))
-		throw new Error((err as { error?: string }).error ?? 'Planning failed')
-	}
-	return (await response.json()) as PlanResult
-}
-
-export interface ClarifyRequest {
-	prompt: string
-	brandText: string
-	outputType: OutputType
-}
-
-export interface ClarifyResult {
-	questions: string[]
+	return post('/api/plan', params, 'Planning failed')
 }
 
 /**
@@ -104,14 +40,18 @@ export interface ClarifyResult {
  * is already clear.
  */
 export async function apiClarify(params: ClarifyRequest): Promise<ClarifyResult> {
-	const response = await fetch('/api/clarify', {
+	return post('/api/clarify', params, 'Could not get questions')
+}
+
+async function post<Result>(path: string, params: unknown, fallbackError: string): Promise<Result> {
+	const response = await fetch(path, {
 		method: 'POST',
 		headers: { 'Content-Type': 'application/json' },
 		body: JSON.stringify(params),
 	})
 	if (!response.ok) {
 		const err = await response.json().catch(() => ({ error: response.statusText }))
-		throw new Error((err as { error?: string }).error ?? 'Could not get questions')
+		throw new Error((err as { error?: string }).error ?? fallbackError)
 	}
-	return (await response.json()) as ClarifyResult
+	return (await response.json()) as Result
 }
