+++ apps/examples/marketing-assets/tsconfig.json
@@ -17,7 +17,7 @@
 		"noFallthroughCasesInSwitch": true,
 		"types": ["vite/client", "./worker-configuration.d.ts"]
 	},
-	"include": ["src", "worker", "vite.config.ts"],
+	"include": ["src", "worker", "shared", "vite.config.ts"],
 	"references": [
 		{ "path": "../../../packages/sync" },
 		{ "path": "../../../packages/sync-core" },
